+++ components/collector/src/source_collectors/jira/test_cases.py
@@ -1,0 +1,7 @@
+"""Jira test cases collector."""
+
+from .issues import JiraIssues
+
+
+class JiraTestCases(JiraIssues):
+    """Jira collector for test cases."""