+++ components/collector/src/source_collectors/testng/test_cases.py
@@ -1,0 +1,7 @@
+"""TestNG test cases collector."""
+
+from .tests import TestNGTests
+
+
+class TestNGTestCases(TestNGTests):
+    """Collector for TestNG test cases."""