+++ components/server/src/data_model/sources/testng.py
@@ -5,7 +5,20 @@ from ..meta.source import Source
 from ..parameters import access_parameters, TestResult
 
 
-ALL_TESTNG_METRICS = ["source_up_to_dateness", "tests"]
+ALL_TESTNG_METRICS = ["source_up_to_dateness", "test_cases", "tests"]
+
+TEST_ENTITIES = dict(
+    name="test",
+    attributes=[
+        dict(name="Class name"),
+        dict(name="Test method", key="name"),
+        dict(name="Description"),
+        dict(
+            name="Test result",
+            color=dict(failed=Color.NEGATIVE, passed=Color.POSITIVE, skipped=Color.WARNING),
+        ),
+    ],
+)
 
 TESTNG = Source(
     name="TestNG",
@@ -15,18 +28,5 @@ TESTNG = Source(
         test_result=TestResult(values=["failed", "ignored", "passed", "skipped"]),
         **access_parameters(ALL_TESTNG_METRICS, source_type="TestNG report", source_type_format="XML")
     ),
-    entities=dict(
-        tests=dict(
-            name="test",
-            attributes=[
-                dict(name="Class name"),
-                dict(name="Test method", key="name"),
-                dict(name="Description"),
-                dict(
-                    name="Test result",
-                    color=dict(failed=Color.NEGATIVE, passed=Color.POSITIVE, skipped=Color.WARNING),
-                ),
-            ],
-        )
-    ),
+    entities=dict(tests=TEST_ENTITIES, test_cases=TEST_ENTITIES),
 )